DAOS-5065 csum: Server Checksum Fetch Logic Rework
This patch reworks the server side logic for retrieving
checksums on a fetch. It is more performant for larger
extents and has been simplified to be easier to read. Now
instead of iterating over each chunk and determining what
to do for each, the logic iterates biovs. If the biov
is a hole, then it skips it, otherwise it first considers
the prefix/suffix of a biov and if current index is in the
same chunk as the previx or suffix a new checksum is needed.
The logic also considers if there is a different non
hole biov in the same chunk. If a new checksum is needed
then the logic will continue iterating over biovs, using
the data to calculate the new checksum until the end of
the chunk or request is reached.

If a new checksum is not needed, then it will copy as much
of the biov as possible which would include all but
possibly the last chunk of the biov.

Also ...
- Updated checksum_timing utility
- For the srv checksum tests, changed how checksums are
  verified. Instead of checking specific csum results, now
  just looking to see if new checksums are created or copied
  from what's stored. This makes it easier to read (it's a
  string now) and verify the server checksum logic.
- Added several new server checksum tests
- Added more testing to checksum functional tests.
- Setup daos_checksum tests to test all checksum types
  with an env variable set. Don't want it to run that way in
  CI as it can be slow.
- Fixed bug where evtree function to update biov's with
  prefix/suffix and adjust csums wasn't taking record len
  into account. Was also double offsetting the
  checksum (evtree.c and vos_io.c)
- Updated data_integrity and readme documentation

Signed-off-by: Ryon Jensen <ryon.jensen@intel.com>